+++ internal/elasticsearch/index/ilm.go
@@ -273,6 +273,31 @@ var supportedActions = map[string]*schema.Schema{
 					Type:        schema.TypeString,
 					Optional:    true,
 				},
+				"min_age": {
+					Description: "Prevents rollover until after the minimum elapsed time from index creation is reached. Supported from Elasticsearch version **8.4**",
+					Type:        schema.TypeString,
+					Optional:    true,
+				},
+				"min_docs": {
+					Description: "Prevents rollover until after the specified minimum number of documents is reached. Supported from Elasticsearch version **8.4**",
+					Type:        schema.TypeInt,
+					Optional:    true,
+				},
+				"min_size": {
+					Description: "Prevents rollover until the index reaches a certain size.",
+					Type:        schema.TypeString,
+					Optional:    true,
+				},
+				"min_primary_shard_size": {
+					Description: "Prevents rollover until the largest primary shard in the index reaches a certain size. Supported from Elasticsearch version **8.4**",
+					Type:        schema.TypeString,
+					Optional:    true,
+				},
+				"min_primary_shard_docs": {
+					Description: "Prevents rollover until the largest primary shard in the index reaches a certain number of documents. Supported from Elasticsearch version **8.4**",
+					Type:        schema.TypeInt,
+					Optional:    true,
+				},
 			},
 		},
 	},
@@ -478,7 +503,7 @@ func expandPhase(p map[string]interface{}, serverVersion *version.Version) (*mod
 					}
 				}
 			case "rollover":
-				actions[actionName], diags = expandAction(a, serverVersion, "max_age", "max_docs", "max_size", "max_primary_shard_size")
+				actions[actionName], diags = expandAction(a, serverVersion, "max_age", "max_docs", "max_size", "max_primary_shard_size", "min_age", "min_docs", "min_size", "min_primary_shard_size", "min_primary_shard_docs")
 			case "searchable_snapshot":
 				actions[actionName], diags = expandAction(a, serverVersion, "snapshot_repository", "force_merge_index")
 			case "set_priority":
@@ -509,14 +534,20 @@ func expandPhase(p map[string]interface{}, serverVersion *version.Version) (*mod
 	return &phase, diags
 }
 
+var RolloverMinConditionsMinSupportedVersion = version.Must(version.NewVersion("8.4.0"))
 var ilmActionSettingOptions = map[string]struct {
 	skipEmptyCheck bool
 	def            interface{}
 	minVersion     *version.Version
 }{
-	"number_of_replicas":    {skipEmptyCheck: true},
-	"total_shards_per_node": {skipEmptyCheck: true, def: -1, minVersion: version.Must(version.NewVersion("7.16.0"))},
-	"priority":              {skipEmptyCheck: true},
+	"number_of_replicas":     {skipEmptyCheck: true},
+	"total_shards_per_node":  {skipEmptyCheck: true, def: -1, minVersion: version.Must(version.NewVersion("7.16.0"))},
+	"priority":               {skipEmptyCheck: true},
+	"min_age":                {def: "", minVersion: RolloverMinConditionsMinSupportedVersion},
+	"min_docs":               {def: 0, minVersion: RolloverMinConditionsMinSupportedVersion},
+	"min_size":               {def: "", minVersion: RolloverMinConditionsMinSupportedVersion},
+	"min_primary_shard_size": {def: "", minVersion: RolloverMinConditionsMinSupportedVersion},
+	"min_primary_shard_docs": {def: 0, minVersion: RolloverMinConditionsMinSupportedVersion},
 }
 
 func expandAction(a []interface{}, serverVersion *version.Version, settings ...string) (map[string]interface{}, diag.Diagnostics) {